requirements.txt
torch
transformers==4.40.2
sentencepiece
accelerate
openai-whisper
